+++ examples/ls-glob.py
@@ -8,14 +8,17 @@ handling. For an involved example see the "do_ls" method in "bin/mantash"
 for glob handling.
 
 Usage:
-    python ls-glob.py <manta-path-with-file-glob>
+    python ls-glob.py [-v] <manta-path-with-file-glob>
 
 Example:
     $ python ls-glob.py /trent.mick/stor/tmp/*.m4a
     /trent.mick/stor/tmp/blue.m4a
     /trent.mick/stor/tmp/excursion.m4a
+
+Use '-v' option for verbose output.
 """
 
+import logging
 import os
 from pprint import pprint
 import sys
@@ -27,7 +30,7 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 import manta
 
 
-def get_client():
+def get_client(verbose=False):
     MANTA_USER = os.environ['MANTA_USER']
     MANTA_URL = os.environ['MANTA_URL']
     MANTA_TLS_INSECURE = bool(os.environ.get('MANTA_TLS_INSECURE', False))
@@ -40,15 +43,21 @@ def get_client():
     client = manta.MantaClient(url=MANTA_URL,
         account=MANTA_USER,
         signer=signer,
-        # Uncomment this for verbose client output for test run.
-        #verbose=True,
+        verbose=verbose,
         disable_ssl_certificate_validation=MANTA_TLS_INSECURE)
     return client
 
 
 #---- mainline
 
-client = get_client()
+logging.basicConfig()
+if '-v' in sys.argv:
+    sys.argv.remove('-v')
+    verbose = True
+else:
+    verbose = False
+
+client = get_client(verbose)
 if len(sys.argv) < 2:
     sys.stderr.write('ls-glob.py error: not enough arguments\n\n')
     sys.stderr.write('usage: python ls-glob.py <manta-path-with-file-glob>\n')